+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Dagon 0.33.0 - TBD
+------------------
+- **Core**
+  - New config options: `logLevel`, `windowResizable`, `windowX`, `windowY`, `vsync`, `glDebugOutput`.
+
 Dagon 0.32.1 - 20 Oct, 2025
 ---------------------------
 - **UI**